Move UI-free query_patterns_metrics under tui/views/providers
