+++ src/interpreter/context.rs
@@ -70,7 +70,8 @@ pub struct Context {
 
     pub queries_filter: Arc<Mutex<String>>,
     pub queries_limit: Arc<Mutex<u64>>,
-    pub query_patterns_metric: &'static crate::view::providers::query_patterns_metrics::Metric,
+    pub query_patterns_metric:
+        &'static crate::tui::views::providers::query_patterns_metrics::Metric,
 
     pub debug_metrics: Arc<DebugMetrics>,
 }
@@ -93,7 +94,7 @@ impl Context {
         let queries_filter = Arc::new(Mutex::new(String::new()));
         let queries_limit = Arc::new(Mutex::new(options.view.queries_limit));
         let query_patterns_metric =
-            crate::view::providers::query_patterns_metrics::default_metric();
+            crate::tui::views::providers::query_patterns_metrics::default_metric();
 
         // Metrics are always collected; display is toggled with `!`. The refresh thread
         // sleeps when hidden, so this is free when unused.
