+++ src/tui/views/mod.rs
@@ -1,5 +1,14 @@
 pub mod log_store;
+pub mod log_view;
+pub mod providers;
+pub mod queries_view;
+pub mod query_view;
 pub mod search_history;
+pub mod settings_view;
+pub mod sql_query_view;
+pub mod summary_view;
+pub mod table_view;
+pub mod text_log_view;
 
 pub use log_store::{LogEntry, LogStore};
 pub use search_history::SearchHistory;
