+++ src/tui/views/providers/mod.rs
@@ -1,0 +1,1 @@
+pub mod query_patterns_metrics;
