+++ src/tui/views/providers/query_patterns_metrics.rs
