+++ src/view/providers/mod.rs
@@ -15,7 +15,7 @@ mod object_storage_queue;
 pub mod part_log;
 mod queries;
 pub mod query_patterns;
-pub mod query_patterns_metrics;
+pub use crate::tui::views::providers::query_patterns_metrics;
 mod replicas;
 mod replicated_fetches;
 mod replication_queue;
